% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{cts_to_vsd}
\alias{cts_to_vsd}
\title{Generate vsd object from counts and metadata}
\usage{
cts_to_vsd(counts, metadata)
}
\arguments{
\item{counts}{A matrix - the RNA-seq count matrix}

\item{metadata}{A metadata - the metadata matrix}
}
\value{
A vsd object
}
\description{
Generate vsd object from counts and metadata
}
