% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{filter_mut_table}
\alias{filter_mut_table}
\title{Filter raw mutation tables}
\usage{
filter_mut_table(
  table,
  t_cov = 15,
  n_cov = 8,
  t_frac = 0.02,
  n_frac = 0.01,
  pop_freq = 0.01
)
}
\arguments{
\item{table}{A dataframe - the mutation table downloaded from server}

\item{t_cov}{An integer - the minimum coverage for tumor sample}

\item{n_cov}{An integer - the minimum coverage for normal sample}

\item{t_frac}{A double - the minimum allele fraction in tumor sample}

\item{n_frac}{A double - the maximum allele fraction in normal sample}

\item{pop_freq}{A double - the maximum population frequncy}
}
\value{
A dataframe - the table with added columns
}
\description{
Filter raw mutation tables
}
