% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneInfo}
\alias{geneInfo}
\title{Hg19 gene start-end coordinates for CNV analysis}
\format{
A dataframe

\describe{
\item{chrom}{Chromosome number.}
\item{start}{The starting coordinate.}
\item{end}{The ending coordinate}
\item{geneid}{Gene ID}
\item{genename}{Gene Symbol}
}
}
\usage{
geneInfo
}
\description{
A dataframe of hg19 gene coordinates
}
\examples{
head(geneInfo)
}
\keyword{datasets}
