% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_colors}
\alias{get_all_colors}
\title{Get all colors in an RColorBrewer palette}
\usage{
get_all_colors(pal)
}
\arguments{
\item{pal}{A string - the name of \code{RColorBrewer} palette}
}
\value{
An string vector
}
\description{
Get all colors in an RColorBrewer palette
}
