% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get_mtx_dds}
\alias{get_mtx_dds}
\title{Get sample-gene matrix from DESeqDataSet object}
\usage{
get_mtx_dds(dds, genes, raw = F)
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{genes}{A string vector - list of genes}

\item{raw}{A logical - whether to get raw counts data. Default value is \code{FALSE}.}
}
\value{
A matrix
}
\description{
Get sample-gene matrix from DESeqDataSet object
}
