% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get_nm_count_dds}
\alias{get_nm_count_dds}
\title{Get sample-gene matrix from DESeqDataSet object}
\usage{
get_nm_count_dds(dds, genes, var)
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{genes}{A string vector - list of genes}

\item{var}{A string - the name of the variable matching metadata columns}
}
\value{
A tibble
}
\description{
Get sample-gene matrix from DESeqDataSet object
}
