% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{get_top_genes}
\alias{get_top_genes}
\title{Get top n genes from DESeq2Results object}
\usage{
get_top_genes(res, topn = 50)
}
\arguments{
\item{res}{A DESeq2Results object}

\item{topn}{An integer - the number of top genes to be selected. Default value is \code{50}.}
}
\value{
A string vector
}
\description{
Get top n genes from DESeq2Results object
}
