% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{htseq_to_mtx}
\alias{htseq_to_mtx}
\title{Generate raw count matrix from HTSeq count files}
\usage{
htseq_to_mtx(files)
}
\arguments{
\item{files}{A string vector or a dataframe - the file paths or file connections of HTSeq count files}
}
\value{
A matrix
}
\description{
Generate raw count matrix from HTSeq count files
}
