% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{log2ratio_to_segment}
\alias{log2ratio_to_segment}
\title{Get segmentation data from raw.txt files generated by Nexus}
\usage{
log2ratio_to_segment(file, sample_name = NULL, pattern = NULL)
}
\arguments{
\item{file}{A string vector - list of raw.txt file names}

\item{sample_name}{A string vector - matching the names of \code{file}. Default value is \code{NULL}.}

\item{pattern}{A string - the pattern used to generate sample names if \code{sample_name} is not specified.}
}
\value{
A dataframe of segmentation data
}
\description{
Get segmentation data from raw.txt files generated by Nexus
}
