% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{mtx_name_match}
\alias{mtx_name_match}
\title{Matching file names with sample names from metadata}
\usage{
mtx_name_match(mtx, metadata, sample_col, file_col, cutoff)
}
\arguments{
\item{mtx}{A matrix - the raw gene count matrix}

\item{metadata}{A dataframe - the metadata}

\item{sample_col}{A string - the metadata column with sample names}

\item{file_col}{A string - the metadata column with file names}

\item{cutoff}{An integer - the cutoff for filtering low-expressed genes. Genes with total counts less than this cutoff will be excluded.}
}
\value{
A matrix
}
\description{
Matching file names with sample names from metadata
}
