% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{mtx_rescale}
\alias{mtx_rescale}
\title{Rescale RNA-seq sample-gene matrix}
\usage{
mtx_rescale(mtx)
}
\arguments{
\item{mtx}{A matrix - the sample-gene matrix}
}
\value{
A matrix
}
\description{
Rescale RNA-seq sample-gene matrix
}
