% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mut_add_column}
\alias{mut_add_column}
\title{Add columns to the raw mutation tables}
\usage{
mut_add_column(table, pattern)
}
\arguments{
\item{table}{A dataframe - the mutation table downloaded from server}

\item{pattern}{A string - the pattern to extract sample names}
}
\value{
A dataframe - the table with added columns
}
\description{
Add columns to the raw mutation tables
}
