% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{num_colors}
\alias{num_colors}
\title{Number of colors in the RColorBrewer palette}
\usage{
num_colors(pal)
}
\arguments{
\item{pal}{A string - the name of \code{RColorBrewer} palette}
}
\value{
An integer
}
\description{
Number of colors in the RColorBrewer palette
}
