% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_deseq_volcano}
\alias{plot_deseq_volcano}
\title{Volcano plot from DESeq2Results object}
\usage{
plot_deseq_volcano(
  res,
  p_co = 0.05,
  lfc_co = 2,
  p_plot_lim = 5,
  lfc_plot_lim = 5
)
}
\arguments{
\item{res}{A DESeq2Results object}

\item{p_co}{A double - the cutoff of adjusted p-value. Default values is \code{0.05}.}

\item{lfc_co}{A double - the cutoff of log2 fold change. Default values is \code{2}.}

\item{p_plot_lim}{A double - the y-limit of -log10 adjusted p-value. Default value is \code{5}.}

\item{lfc_plot_lim}{A double - the x-limit of log2 fold change plot. Default value is \code{5}.}
}
\value{
A ggplot2 plot
}
\description{
Volcano plot from DESeq2Results object
}
