% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_gene_boxplot}
\alias{plot_gene_boxplot}
\title{Boxplot from DESeqDataSet object}
\usage{
plot_gene_boxplot(dds, genes, var, pal = options("xmcutil.tripalette")[[1]][1])
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{genes}{A string vector - list of genes}

\item{var}{A string - the name of the variable matching metadata columns}

\item{pal}{A string - palette name of \code{RColorBrewer}. Default value takes the palettes set up in \code{options('xmc.tripalette')}.}
}
\value{
A heatmap
}
\description{
Boxplot from DESeqDataSet object
}
