% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_pca_df}
\alias{plot_pca_df}
\title{PCA plot from PCA coordinates}
\usage{
plot_pca_df(df, var)
}
\arguments{
\item{df}{A dataframe - the PCA coordinates of RNA-seq samples}

\item{var}{A string - the name of the variable matching metadata columns}
}
\value{
A ggplot2 plot
}
\description{
PCA plot from PCA coordinates
}
