% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_pca_vsd}
\alias{plot_pca_vsd}
\title{PCA plot from vsd object}
\usage{
plot_pca_vsd(vsd, var, pal)
}
\arguments{
\item{vsd}{A vsd object}

\item{var}{A string - the name of the variable matching metadata columns}

\item{pal}{A string - palette name of \code{RColorBrewer}}
}
\value{
A ggplot2 plot
}
\description{
PCA plot from vsd object
}
