% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_rna_pca}
\alias{plot_rna_pca}
\title{PCA plot from PCA coordinates}
\usage{
plot_rna_pca(counts, metadata, vars)
}
\arguments{
\item{counts}{A matrix - the RNA-seq count matrix}

\item{metadata}{A metadata - the metadata matrix}

\item{vars}{A string vector - the vector of varibles to plot}
}
\value{
A list ggplot2 plots
}
\description{
PCA plot from PCA coordinates
}
