% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_sample_gene_mtx}
\alias{plot_sample_gene_mtx}
\title{Sample-gene matrix heatmap from DESeqDataSet object}
\usage{
plot_sample_gene_mtx(dds, genes, pal, dir)
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{genes}{A string vector - list of genes}

\item{pal}{A string - palette name of \code{RColorBrewer}}

\item{dir}{An integer - \code{1} or \code{-1}, to adjust the direction of colors.}
}
\value{
A heatmap
}
\description{
Sample-gene matrix heatmap from DESeqDataSet object
}
