% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{res_add_shape}
\alias{res_add_shape}
\title{Add shape column for log2 fold change limit}
\usage{
res_add_shape(res, lfc_plot_lim)
}
\arguments{
\item{res}{A tibble}

\item{lfc_plot_lim}{A double - the x-limit of log2 fold change plot}
}
\value{
A tibble
}
\description{
Add shape column for log2 fold change limit
}
