% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{res_to_tibble}
\alias{res_to_tibble}
\title{Transform DESeq2Results object to tibble}
\usage{
res_to_tibble(res, p_co, lfc_co)
}
\arguments{
\item{res}{A DESeq2Results object}

\item{p_co}{A double - the cutoff of adjusted p-value}

\item{lfc_co}{A double - the cutoff of log2 fold change}
}
\value{
A tibble
}
\description{
Transform DESeq2Results object to tibble
}
