% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{xmc_constants}
\alias{xmc_constants}
\title{Helper function to generate constants used by other functions}
\usage{
xmc_constants()
}
\description{
Helper function to generate constants used by other functions
}
