% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_deseq_gsea_list}
\alias{plot_deseq_gsea_list}
\title{Dotplot from a list of GSEA results}
\usage{
plot_deseq_gsea_list(gsea_list, p_co)
}
\arguments{
\item{gsea_list}{A list of tibble of GSEA results}

\item{p_co}{A double - the cutoff of adjusted p-value}
}
\value{
A ggplot2 plot
}
\description{
Dotplot from a list of GSEA results
}
