% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_heatmap_vsd}
\alias{plot_heatmap_vsd}
\title{Heatmap from vsd object}
\usage{
plot_heatmap_vsd(vsd, var, pal = xmc_constants()$palette[2], dir = 1)
}
\arguments{
\item{vsd}{A vsd object}

\item{var}{A string - the name of the variable matching metadata columns}

\item{pal}{A string - palette name of \code{RColorBrewer}. Default value takes the palette set up in \code{xmc_constants()}.}

\item{dir}{An integer - \code{0} or \code{1}, to adjust the direction of colors. Default value is \code{1}.}
}
\value{
A heatmap
}
\description{
Heatmap from vsd object
}
