% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{plot_pca_vsd}
\alias{plot_pca_vsd}
\title{PCA plot from vsd object}
\usage{
plot_pca_vsd(vsd, var, pal = NULL, dir = 1)
}
\arguments{
\item{vsd}{A vsd object}

\item{var}{A string - the name of the variable matching metadata columns}

\item{pal}{A string - palette name of \code{RColorBrewer}. Default value is \code{NULL}, taking the palettes set up in \code{xmc_constants()}.}

\item{dir}{An integer - \code{1} or \code{-1}, to adjust the direction of colors. Default value is \code{1}.}
}
\value{
A ggplot2 plot
}
\description{
PCA plot from vsd object
}
