% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{segment_to_cnv}
\alias{segment_to_cnv}
\title{Get CNV status from segmentation data}
\usage{
segment_to_cnv(df, gene_list, return_num = FALSE, cutoff = c(-0.3, 0.3))
}
\arguments{
\item{df}{A dataframe - segmentation data}

\item{gene_list}{A string vector - list of human gene symbols.}

\item{return_num}{A logical - whether numeric value is returned or not. Default is \code{FALSE}.}

\item{cutoff}{A numeric vector of length 2}
}
\value{
A dataframe of CNV status
}
\description{
Get CNV status from segmentation data
}
