% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{segment_to_cnv}
\alias{segment_to_cnv}
\title{Get CNV status from segmentation data}
\usage{
segment_to_cnv(df, gene_list = NULL, as_matrix = FALSE, ...)
}
\arguments{
\item{df}{A dataframe - segmentation data}

\item{gene_list}{A string vector - list of human gene symbols. Default value is \code{NULL}, indicating the whole genome.}

\item{as_matrix}{A logical value - if the result table is cast to matrix format. Default value is \code{FALSE}.}

\item{...}{Additional parameters passed to \code{test_cnv()}}
}
\value{
A dataframe of CNV status
}
\description{
Get CNV status from segmentation data
}
