% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.R
\name{test_cnv}
\alias{test_cnv}
\title{Look up corresponding values from a lookup table}
\usage{
test_cnv(values, cutoff = c(-0.3, 0.3))
}
\arguments{
\item{values}{A numeric vector - the log2Ratio values for copy number changes.}

\item{cutoff}{A numeric vector of length 2 - the cutoff for calling amplifications or deletions}
}
\value{
A numeric vector of the same length with \code{values}
}
\description{
Look up corresponding values from a lookup table
}
